fix(menu) : fix some bug and improve design css

+++ Page/menu.php
@@ -44,8 +44,7 @@ if (isset($_GET['delete'])) {
     <title>รายการอาหาร</title>
 
     <!-- Bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9" crossorigin="anonymous" />
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9" crossorigin="anonymous" />
     <!-- CSS -->
     <link rel="stylesheet" href="../css/index.css" />
     <link rel="stylesheet" href="../css/menu.css" />
@@ -62,8 +61,7 @@ if (isset($_GET['delete'])) {
             </a>
 
             <!-- Add the hamburger menu button -->
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav"
-                aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
 
@@ -74,6 +72,9 @@ if (isset($_GET['delete'])) {
                         <p class="text-username">ร้าน
                             <?php echo htmlspecialchars($username); ?>
                         </p>
+                    </li>
+                    <li class="nav-item">
+                        <a  href="../page/result/result.php" class="btn btn-store-atnav" style="gap:8px;"><i class="bi bi-receipt"></i>ยอดคำสั่งซื้อ</a>
                     </li>
                     <li class="nav-item">
                         <a href="../service/logout.php" class="btn btn-order-atnav">ออกจากระบบ</a>
@@ -149,8 +150,7 @@ if (isset($_GET['delete'])) {
                             <?php } ?>
                         <?php } ?>
                         <!-- Button to generate QR Code -->
-                        <button type="button" class="btn btn-qrcode-gen" id="generate-qrcode-btn"
-                            onclick="generateQRCode()">สร้าง QRCODE</button>
+                        <button type="button" class="btn btn-qrcode-gen" id="generate-qrcode-btn" onclick="generateQRCode()">สร้าง QRCODE</button>
                     </div>
                 </div>
             </div>
@@ -166,11 +166,9 @@ if (isset($_GET['delete'])) {
                     </h4>
                 </div>
                 <div class="col-md-6 d-flex justify-content-end" style="gap: 8px;">
-                    <button type="button" class="btn btn-add-menu" data-bs-toggle="modal"
-                        data-bs-target="#qrcodemodal"><i class="bi bi-qr-code-scan"></i>แสดง QRCODE</button>
+                    <button type="button" class="btn btn-add-menu" data-bs-toggle="modal" data-bs-target="#qrcodemodal"><i class="bi bi-qr-code-scan"></i>แสดง QRCODE</button>
                     <button type="button" class="btn btn-add-menu" onclick="addtable()"><i class="bi bi-plus-circle-fill"></i>เพิ่มโต๊ะอาหาร</button>
-                    <button type="button" class="btn btn-add-menu" data-bs-toggle="modal" data-bs-target="#foodmodal"><i
-                            class="bi bi-plus-circle-fill"></i>เพิ่มรายการอาหาร</button>
+                    <button type="button" class="btn btn-add-menu" data-bs-toggle="modal" data-bs-target="#foodmodal"><i class="bi bi-plus-circle-fill"></i>เพิ่มรายการอาหาร</button>
                 </div>
             </div>
             <hr>
@@ -211,7 +209,7 @@ if (isset($_GET['delete'])) {
                                 echo "<p><td colspan='6' class='text-center'>No data available</td></p>";
                             } else {
                                 foreach ($food as $food) {
-                                    ?>
+                            ?>
                                     <tr>
                                         <th scope="row">
                                             <?php echo $food['food_ID']; ?>
@@ -223,28 +221,21 @@ if (isset($_GET['delete'])) {
                                             <?php echo $food['price']; ?>
                                         </td>
 
-                                        <td width="250px"><img class="rounded" width="100%"
-                                                src="../upload/<?php echo $food['img']; ?>" alt=""></td>
+                                        <td width="250px"><img class="rounded" width="100%" src="../upload/<?php echo $food['img']; ?>" alt=""></td>
                                         <td class="action-btn-layout column">
-                                            <a href="../Page/edit_menu.php?id=<?php echo $food['food_ID']; ?>"
-                                                class="btn btn-edit">แก้ไข</a>
-                                            <a onclick="return confirm('Are you sure you want to delete?');"
-                                                href="?delete=<?php echo $food['food_ID']; ?>" class="btn btn-delete">ลบ</a>
+                                            <a href="../Page/edit_menu.php?id=<?php echo $food['food_ID']; ?>" class="btn btn-edit">แก้ไข</a>
+                                            <a onclick="return confirm('Are you sure you want to delete?');" href="?delete=<?php echo $food['food_ID']; ?>" class="btn btn-delete">ลบ</a>
                                         </td>
                                     </tr>
-                                <?php }
+                            <?php }
                             } ?>
                         </tbody>
                     </table>
+                    </div>
                 </div>
-            </div>
         </div>
-        <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js"
-            integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r"
-            crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.min.js"
-            integrity="sha384-Rx+T1VzGupg4BHQYs2gCW9It+akI2MM/mndMCy36UVfodzcJcF0GGLxZIzObiEfa"
-            crossorigin="anonymous"></script>
+        <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js" integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r" crossorigin="anonymous"></script>
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.min.js" integrity="sha384-Rx+T1VzGupg4BHQYs2gCW9It+akI2MM/mndMCy36UVfodzcJcF0GGLxZIzObiEfa" crossorigin="anonymous"></script>
         <script>
             let imgInput = document.getElementById('imgInput');
             let previewImg = document.getElementById('previewImg');
